+++ release/missRanger/R/generateNA.R
@@ -5,7 +5,7 @@
 #' @author Michael Mayer
 #' 
 #' @param x A vector, matrix or \code{data.frame}.
-#' @param p Proportion of missing values to add to \code{x}. If \code{x} is a \code{data.frame}, each column will receive the same amount of missing values. Use a vector valued \code{p} to apply different proportions of missing values to each column.
+#' @param p Proportion of missing values to add to \code{x}. In case \code{x} is a \code{data.frame}, \code{p} can also be a vector of probabilities per column or a named vector (see examples).
 #' @param seed An integer seed.
 #'
 #' @return \code{x} with missing values.
@@ -13,9 +13,14 @@
 #' @export
 #'
 #' @examples 
+#' generateNA(1:10, p = 0.5, seed = 3345)
+#' generateNA(rep(Sys.Date(), 10))
+#' generateNA(cbind(1:10, 10:1), p = 0.2)
+#' head(generateNA(iris))
 #' head(generateNA(iris, p = 0.2))
 #' head(generateNA(iris, p = c(0, 1, 0.5, 0.5, 0.5)))
-#' generateNA(1:10, p = 0.5, seed = 3345)
+#' head(generateNA(iris, p = c(Sepal.Length = 1)))
+#' head(generateNA(iris, p = c(Species = 0.2, Sepal.Length = 0.5)))
 generateNA <- function(x, p = 0.1, seed = NULL) {
   stopifnot(p >= 0, p <= 1, is.atomic(x) || is.data.frame(x))
   
@@ -29,11 +34,14 @@ generateNA <- function(x, p = 0.1, seed = NULL) {
     z
   }
   
+  # vector or matrix
   if (is.atomic(x)) {
     return(generateNaVec(x, p))
   } 
   
-  x[] <- Map(generateNaVec, x, p)
- 
+  # data frame
+  v <- if (is.null(names(p))) names(x) else intersect(names(p), names(x))
+  x[, v] <- Map(generateNaVec, x[, v, drop = FALSE], p)
+
   x
 }